Stop being stubborn and admit that it's okay to use inheritance.
