+++ java/client/src/org/openqa/selenium/remote/BUCK
@@ -101,6 +101,8 @@ java_library(name = 'remote-lib',
     'http/HttpMethod.java',
     'http/HttpRequest.java',
     'http/HttpResponse.java',
+    'http/AbstractHttpCommandCodec.java',
+    'http/AbstractHttpResponseCodec.java',
     'http/JsonHttpCommandCodec.java',
     'http/JsonHttpResponseCodec.java',
     'http/W3CHttpCommandCodec.java',
